Several changes bundled together due to editor problems
- Switched to ant compilation
- Moved source into src/ folder
- Switching project to use Short[] for the images

+++ src/SVQ.java
@@ -21,18 +21,18 @@ public class SVQ {
         }
     }
 
-    public void checkLengths(double[] a, double[] b){
+    public void checkLengths(short[] a, short[] b){
         checkLengths(a, b.length);
     }
 
-    public void checkLengths(double[] a, int bLen){
+    public void checkLengths(short[] a, int bLen){
         if (a.length != bLen) {
             throw new RuntimeException(
                 "Vectors lengths don't match.");
         }
     }
 
-    public double[] similarities(double[] vec) {
+    public double[] similarities(short[] vec) {
         checkLengths(vec, imgsize);
         double[] ret = new double[ncentr];
         // System.out.print("Simils: ");
@@ -55,9 +55,9 @@ public class SVQ {
         return ret;
     }
 
-    public double[] code(double[] vec) {
-        double[] ret = new double[ncentr];
-        Arrays.fill(ret, 0d);
+    public short[] code(short[] vec) {
+        short[] ret = new short[ncentr];
+        Arrays.fill(ret, (short)0);
 
         double[] mwi = maxWithIndex(similarities(vec));
         // System.out.println("Max: "+mwi[0]+" - idx: "+mwi[1]);
@@ -66,14 +66,10 @@ public class SVQ {
         return ret;
     }
 
-    public double[] reconstruct(double[] code) {
+    public short[] reconstruct(short[] code) {
         checkLengths(code, ncentr);
-        double[] ret = new double[imgsize];
-        Arrays.fill(ret, 0d);
-        // for (int i=0; i<code.length; i++) {
-        //     System.out.print(code[i] + " ");
-        // }
-        // System.out.println();
+        short[] ret = new short[imgsize];
+        Arrays.fill(ret, (short)0);
 
         for (int imgPos=0; imgPos<imgsize; imgPos++) {
             for (int cIdx=0; cIdx<ncentr; cIdx++) {
@@ -83,19 +79,19 @@ public class SVQ {
         return ret; //(new BMPLoader("","")).rescale(ret);
     }
 
-    public double[] reconstructionError(double[] orig, double[] reconstr) {
+    public short[] reconstructionError(short[] orig, short[] reconstr) {
         checkLengths(orig, reconstr);
-        double[] ret = new double[orig.length];
+        short[] ret = new short[orig.length];
         // simple difference - possibly use squared error instead
         for (int i=0; i<orig.length; i++) {
-            ret[i] = Math.abs(orig[i]-reconstr[i]);
+            ret[i] = (short)Math.abs(orig[i]-reconstr[i]);
         }
         return ret;
     }
 
-    public double totalReconstructionError(double[] orig, double[] reconstr) {
-        double[] tmp = reconstructionError(orig, reconstr);
-        double ret = 0;
+    public short totalReconstructionError(short[] orig, short[] reconstr) {
+        short[] tmp = reconstructionError(orig, reconstr);
+        short ret = 0;
         // just total it - what a godforsaken language Java is...
         for (int i=0; i<tmp.length; i++) {
             ret += tmp[i];
@@ -104,7 +100,7 @@ public class SVQ {
     }
 
     // train on single image
-    public void train(double[] img) {
+    public void train(short[] img) {
         double[] closestWI = maxWithIndex(similarities(img));
         // here you get the max in closestWI[0] if you need it
         int closestIdx = (int)closestWI[1];
@@ -113,14 +109,14 @@ public class SVQ {
     }
 
     // train on set of images
-    public void train(double[][] imgs) {
+    public void train(short[][] imgs) {
         for (int i=0; i<imgs.length; i++) {
             train(imgs[i]);
         }
     }
 
-    public double[][] getData() {
-        double[][] ret = new double[ncentr][];
+    public short[][] getData() {
+        short[][] ret = new short[ncentr][];
         for (int i=0; i<ncentr; i++) {
             ret[i] = centroids[i].getData();
         }
